Move database managment API functions
