+++ src/dtls_connection_api.c
@@ -26,6 +26,7 @@
 #include <gnutls/dtls.h>
 #include <gnutls/gnutls.h>
 #include <gnutls/x509.h>
+#include "database.h"
 #include "linked_list.h"
 
 #define BUFFER_SIZE 1024
